[SPARK-58592][CORE] Address review comments on MasterStateResponse tests
